Move item logic to item package.
Rename item classes to have Item prefix instead of suffix
