--- Java/src/main/java/com/gildedrose/LegendaryItem.java
@@ -1,7 +1,0 @@
-package com.gildedrose;
-
-public class LegendaryItem implements IItem {
-    @Override
-    public void update() {
-    }
-}
