+++ Java/src/main/java/com/gildedrose/item/IItem.java
@@ -1,4 +1,4 @@
-package com.gildedrose;
+package com.gildedrose.item;
 
 public interface IItem {
     void update();
