+++ Java/src/main/java/com/gildedrose/item/Item.java
@@ -1,4 +1,4 @@
-package com.gildedrose;
+package com.gildedrose.item;
 
 public class Item implements IItem {
