+++ Java/src/main/java/com/gildedrose/item/ItemLegendary.java
@@ -1,0 +1,7 @@
+package com.gildedrose.item;
+
+public class ItemLegendary implements IItem {
+    @Override
+    public void update() {
+    }
+}
